+++ packages/core/src/core.ts
@@ -33,7 +33,7 @@ export type { CoreChartOptions } from "./chart/core-financial-chart";
 export * from "./chart/event-emitter";
 export * from "./chart/formatter";
 export * from "./chart/themes";
-export type { ChartData, TimeRange } from "./chart/types";
+export type { ChartData, ChartDataInput, TimeRange } from "./chart/types";
 export * from "./plugin/extension-theme";
 
 export type {